Implemented UI Design to splash screen, login page and sign up page + updated all colours and themes + moved fragments into fragment folder
