+++ app/src/main/res/layout/fragment_home.xml
@@ -117,8 +117,8 @@
         android:layout_centerHorizontal="true"
         android:translationX="145dp"
         android:translationY="475dp"
-        app:backgroundTint="@color/purple_500"
-        app:rippleColor="@color/purple_500"
+        app:backgroundTint="@color/background_primary"
+        app:rippleColor="@color/background_primary"
         app:tint="@color/white" />
